// src/app/model/claims.ts
export class Claims {
    claimNumber:string;
    MakeModel: string;
    ClaimTitle: string;
    PolicyNumber: string;
    InsuredName: string;
    DateOfEvent: string;
    DateReceieved: string;
    producercode: string;
    EventLocation: string;
    PhoneNumber: string;
    InsuranceCompany: string;
    NumberOfvehicle: string;
    EventInfo: string;
    SupportingDoc:string;
}